text/templates: support multiline actions
Go templates are often used as a scripting or extension language,
describing data and computation, not just outputting it.  For
these use cases, the inability to describe a data structure or
function call across multiple lines can be challenging for
the people doing the scripting or extending.

This PR changes the lexer to treat all Go-defined whitespace
characters equally within the body of an action -- including
the spaces used to indicate trim characters.  In addition,
inline comments are treated as whitespace, allowing documentation
within the body of an expression.