Simplify scanner state and line tracking
Since only the start line of tokens are tracked, the line
count can be updated only after token emit or ignore.  Rune
widths are returned from next() so that backup() can be given
an explicit width, rather than being state-coupled to the
scanner.  Keeping the end position of the input simplifies
peek() and allows it to be inlined.

Being able to save items and emit them later eliminates
the need to munge the scanner position with tricky position
arithmetic.  Similarly tricky/magical position arithmetic
is eliminated by extracting methods for prefix matching,
delimiter searching, etc.  (This will help a lot when
whitespace and comments start being scanned from the
beginning of an action.)